+++ src/index.ts
@@ -15,6 +15,6 @@ export type { ToListConfig, IEnumItems } from './enum-items';
 export type { EnumItemClass, EnumItemOptions } from './enum-item';
 export type { IEnum, EnumInterface, EnumInitOptions } from './enum';
 
-export { KEYS, ITEMS, VALUES, LABELS, IS_ENUM_ITEM, IS_ENUM_ITEMS, IS_ENUM_COLLECTION, defaultLocalize } from './utils';
+export { KEYS, ITEMS, VALUES, LABELS, IS_ENUM_ITEM, IS_ENUM_ITEMS, IS_ENUM, defaultLocalize } from './utils';
 export type { PluginFunc } from './enum';
 export { Enum } from './enum';